trash_tests/pyproject.toml
[tool.poetry]
name = "trash-tests"
version = "0.1.0"
description = ""
authors = ["Arkhip McNutt <ortiz47@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
psycopg2-binary = "^2.9.9"
pytest = "^8.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
